+++ CloneBike/SuperJump.h
@@ -9,14 +9,14 @@ public:
 	~SuperJump();
 	void SetPosition(point<int> point)       { this->currentX = point.x; this->currentY = point.y; }
 	void SetPosition(int xCoord, int yCoord) { currentX = xCoord; currentY = yCoord; }
-	int GetCurrentX()                  const { return currentX; }
+	float GetCurrentX()                  const { return currentX; }
 	int GetItemType()                  const { return itemType; }
 	int GetCurrentLane()               const { return currentY; }
 	void Move();
 private:
 	const int itemType;
 	float currentX;
-	float currentY;
+	int currentY;
 	float currentTime;
 
 };
